Rename engine to opener to match closer
