--- pkg/transport/closer.go
@@ -1,10 +1,0 @@
-package transport
-
-import (
-	"context"
-)
-
-// Closer is the common interface for things that can be closed
-type Closer interface {
-	Close(ctx context.Context) error
-}
