--- pkg/transport/codec.go
@@ -1,29 +1,0 @@
-package transport
-
-//
-//// Codec is the interface for transport codecs to convert between transport
-//// specific payloads and the Message interface.
-//type Codec interface {
-//	Encode(context.Context, event.Event) (Message, error)
-//	Decode(context.Context, Message) (*event.Event, error)
-//}
-//
-//// ErrMessageEncodingUnknown is an error produced when the encoding for an incoming
-//// message can not be understood.
-//type ErrMessageEncodingUnknown struct {
-//	codec     string
-//	transport string
-//}
-//
-//// NewErrMessageEncodingUnknown makes a new ErrMessageEncodingUnknown.
-//func NewErrMessageEncodingUnknown(codec, transport string) *ErrMessageEncodingUnknown {
-//	return &ErrMessageEncodingUnknown{
-//		codec:     codec,
-//		transport: transport,
-//	}
-//}
-//
-//// Error implements error.Error
-//func (e *ErrMessageEncodingUnknown) Error() string {
-//	return fmt.Sprintf("message encoding unknown for %s codec on %s transport", e.codec, e.transport)
-//}
