--- pkg/transport/codec_test.go
@@ -1,9 +1,0 @@
-package transport_test
-
-import (
-	"testing"
-)
-
-func TestCodec(t *testing.T) {
-	// TODO: add a test. This makes coverage count this dir.
-}
